SysUI: Fix reporting of swipe progess in notification panel
SwipHelper only uses half of the child views width for determining
swip progress.  This patch adds a new member variable that can be
changed via a setter.  The default is the constant value that was
used prior to this patch so that it does not affect notifications
that also use the SwipeHelper.