Check whether wallpaper's size has changed
If we are in an activity that shows wallpaper(such as Launcher),
when rotating screen, there will be a short time that wallpaper
shows improperly, maybe showing black in some places.

When the wallpaper's size has changed, wallpaper will redraw itself,
which is not contained in the original rotating logic.

To avoid this, we add check when rotating screen, if wallpaper needs to
be resized, wait for wallpaper redraw itself, then stop freezing screen.