When the application is removed, reset network permission of this uid
Probably not a portal: exception java.net.SocketException: Binding
socket to network 102 failed: EPERM (Operation not permitted)

If there are SYSTEM permissions, then the original is SYSTEM
permissions,
do not update. If other app of this uid have network permissions, the
same
as before do not update, or update permissions. If all app of this uid
do not have any network permission to remove permissions